pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "streamlit>=1.48.1",
]
